move list stories
