+++ packages/react-aria-components/stories/index.stories.tsx
@@ -11,84 +11,15 @@
  */
 
 import {action} from '@storybook/addon-actions';
-import {Button, Checkbox, Dialog, DialogTrigger, DropZone, FileTrigger, Header, Input, Label, Link, ListBox, ListBoxItem, ListBoxProps, Modal, ModalOverlay, Radio, RadioGroup, SearchField, Section, Separator, Switch, Text, TextField, ToggleButton, Toolbar, useDragAndDrop} from 'react-aria-components';
+import {Button, Checkbox, Dialog, DialogTrigger, DropZone, FileTrigger, Input, Label, Link, Modal, ModalOverlay, Radio, RadioGroup, SearchField, Switch, Text, TextField, ToggleButton, Toolbar} from 'react-aria-components';
 import {classNames} from '@react-spectrum/utils';
 import {FocusRing, mergeProps, useButton, useClipboard, useDrag} from 'react-aria';
-import {MyListBoxItem} from './utils';
 import React, {useRef} from 'react';
 import styles from '../example/index.css';
-import {useListData} from 'react-stately';
 
 export default {
   title: 'React Aria Components'
 };
-
-export const ListBoxExample = (args) => (
-  <ListBox className={styles.menu} {...args} aria-label="test listbox">
-    <MyListBoxItem>Foo</MyListBoxItem>
-    <MyListBoxItem>Bar</MyListBoxItem>
-    <MyListBoxItem>Baz</MyListBoxItem>
-    <MyListBoxItem href="http://google.com">Google</MyListBoxItem>
-  </ListBox>
-);
-
-ListBoxExample.story = {
-  args: {
-    selectionMode: 'none',
-    selectionBehavior: 'toggle'
-  },
-  argTypes: {
-    selectionMode: {
-      control: {
-        type: 'radio',
-        options: ['none', 'single', 'multiple']
-      }
-    },
-    selectionBehavior: {
-      control: {
-        type: 'radio',
-        options: ['toggle', 'replace']
-      }
-    }
-  }
-};
-
-// Known accessibility false positive: https://github.com/adobe/react-spectrum/wiki/Known-accessibility-false-positives#listbox
-// also has a aXe landmark error, not sure what it means
-export const ListBoxSections = () => (
-  <ListBox className={styles.menu} selectionMode="multiple" selectionBehavior="replace" aria-label="test listbox with section">
-    <Section className={styles.group}>
-      <Header style={{fontSize: '1.2em'}}>Section 1</Header>
-      <MyListBoxItem>Foo</MyListBoxItem>
-      <MyListBoxItem>Bar</MyListBoxItem>
-      <MyListBoxItem>Baz</MyListBoxItem>
-    </Section>
-    <Separator style={{borderTop: '1px solid gray', margin: '2px 5px'}} />
-    <Section className={styles.group}>
-      <Header style={{fontSize: '1.2em'}}>Section 1</Header>
-      <MyListBoxItem>Foo</MyListBoxItem>
-      <MyListBoxItem>Bar</MyListBoxItem>
-      <MyListBoxItem>Baz</MyListBoxItem>
-    </Section>
-  </ListBox>
-);
-
-export const ListBoxComplex = () => (
-  <ListBox className={styles.menu} selectionMode="multiple" selectionBehavior="replace" aria-label="listbox complex">
-    <MyListBoxItem>
-      <Text slot="label">Item 1</Text>
-      <Text slot="description">Description</Text>
-    </MyListBoxItem>
-    <MyListBoxItem>
-      <Text slot="label">Item 2</Text>
-      <Text slot="description">Description</Text>
-    </MyListBoxItem>
-    <MyListBoxItem>
-      <Text slot="label">Item 3</Text>
-      <Text slot="description">Description</Text>
-    </MyListBoxItem>
-  </ListBox>
-);
 
 function Draggable() {
   let buttonRef = useRef(null);
@@ -318,91 +249,6 @@ export const FileTriggerLinkAllowsMultiple = (props) => (
     <Link>Select a file</Link>
   </FileTrigger>
 );
-
-let albums = [
-  {
-    id: 1,
-    image: 'https://images.unsplash.com/photo-1593958812614-2db6a598c71c?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8Nnx8ZGlzY298ZW58MHx8MHx8fDA%3D&auto=format&fit=crop&w=900&q=60',
-    title: 'Euphoric Echoes',
-    artist: 'Luna Solstice'
-  },
-  {
-    id: 2,
-    image: 'https://images.unsplash.com/photo-1601042879364-f3947d3f9c16?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8M3x8bmVvbnxlbnwwfHwwfHx8MA%3D%3D&auto=format&fit=crop&w=900&q=60',
-    title: 'Neon Dreamscape',
-    artist: 'Electra Skyline'
-  },
-  {
-    id: 3,
-    image: 'https://images.unsplash.com/photo-1528722828814-77b9b83aafb2?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8MTF8fHNwYWNlfGVufDB8fDB8fHww&auto=format&fit=crop&w=900&q=60',
-    title: 'Cosmic Serenade',
-    artist: 'Orion\'s Symphony'
-  },
-  {
-    id: 4,
-    image: 'https://images.unsplash.com/photo-1511379938547-c1f69419868d?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8M3x8bXVzaWN8ZW58MHx8MHx8fDA%3D&auto=format&fit=crop&w=900&q=60',
-    title: 'Melancholy Melodies',
-    artist: 'Violet Mistral'
-  },
-  {
-    id: 5,
-    image: 'https://images.unsplash.com/photo-1608433319511-dfe8ea4cbd3c?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8MTF8fGJlYXR8ZW58MHx8MHx8fDA%3D&auto=format&fit=crop&w=900&q=60',
-    title: 'Rhythmic Illusions',
-    artist: 'Mirage Beats'
-  }
-];
-
-export const ListBoxDnd = (props: ListBoxProps<typeof albums[0]>) => {
-  let list = useListData({
-    initialItems: albums
-  });
-
-  let {dragAndDropHooks} = useDragAndDrop({
-    getItems: (keys) => [...keys].map(key => ({'text/plain': list.getItem(key).title})),
-    onReorder(e) {
-      if (e.target.dropPosition === 'before') {
-        list.moveBefore(e.target.key, e.keys);
-      } else if (e.target.dropPosition === 'after') {
-        list.moveAfter(e.target.key, e.keys);
-      }
-    }
-  });
-
-  return (
-    <ListBox
-      {...props}
-      aria-label="Albums"
-      items={list.items}
-      selectionMode="multiple"
-      dragAndDropHooks={dragAndDropHooks}>
-      {item => (
-        <ListBoxItem>
-          <img src={item.image} alt="" />
-          <Text slot="label">{item.title}</Text>
-          <Text slot="description">{item.artist}</Text>
-        </ListBoxItem>
-      )}
-    </ListBox>
-  );
-};
-
-ListBoxDnd.story = {
-  args: {
-    layout: 'stack',
-    orientation: 'horizontal'
-  },
-  argTypes: {
-    layout: {
-      control: 'radio',
-      options: ['stack', 'grid']
-    },
-    orientation: {
-      control: 'radio',
-      options: ['horizontal', 'vertical']
-    }
-  }
-};
-
 export const RadioGroupExample = () => {
   return (
     <RadioGroup
